pyproject.toml
[project]
name = "basicragwithpython"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.7",
    "fastapi>=0.116.1",
    "google-genai>=1.72.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.0",
    "streamlit>=1.49.0",
    "uvicorn>=0.35.0",
]
